pyproject.toml
[tool.poetry]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = ["thedev-junyoung <thedev.junyoung@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.32.1"
fastapi = "^0.115.6"
py-ulid = "^1.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
